Centralize types
