--- frontend/src/business/models/Genre.ts
@@ -1,8 +1,0 @@
-import { BaseState } from "./State";
-
-export interface Genre {
-  id: string;
-  genre: string;
-}
-
-export type GenreState = BaseState<Genre>;
