+++ frontend/src/business/overmind/Api/state.ts
@@ -1,4 +1,10 @@
-import { ApiState } from "../../models/Api";
+export interface ApiState {
+  response: {
+    statusText?: string;
+    message: string;
+    status: "success" | "error" | undefined;
+  };
+}
 
 export const state: ApiState = {
   response: {
