+++ frontend/src/business/overmind/App/state.ts
@@ -1,4 +1,16 @@
-import { AppState } from "../../models/App";
+export type NavItem = {
+  label: string;
+  subLabel?: string;
+  children?: Array<NavItem>;
+  href?: string;
+};
+
+export interface AppState {
+  isLoading: boolean;
+  publicNavItems: Array<NavItem>;
+  privateNavItems: Array<NavItem>;
+  colors: Array<string>;
+}
 
 export const state: AppState = {
   isLoading: false,
