+++ frontend/src/business/overmind/Rental/state.ts
@@ -1,5 +1,35 @@
 import { derived } from "overmind";
-import { RentalState } from "../../models/Rental";
+import { ApiResponse, BaseState, UITable } from "../../models/State";
+
+export interface Rental {
+  id: string;
+  rental_date: string;
+  return_date: string;
+  title: string;
+  author: string;
+}
+
+export interface RentalStatsDuration {
+  duration: number;
+  total_rentals: number;
+}
+
+export interface RentalState extends BaseState<Rental> {
+  statsDuration: Array<{
+    durationLabel: string;
+    duration: number;
+    number: number;
+  }> | null;
+  statsDurationApi: ApiResponse<Array<RentalStatsDuration>> | null;
+  ui: {
+    table: UITable<
+      Rental,
+      {
+        returned: string;
+      }
+    >;
+  };
+}
 
 export const state: RentalState = {
   getAllApi: null,
