+++ frontend/src/business/overmind/Review/state.ts
@@ -1,6 +1,19 @@
 import { derived } from "overmind";
 import { groupBy } from "../../functions/groupBy";
-import { ReviewState } from "../../models/Review";
+import { ApiResponse } from "../../models/State";
+
+export interface Review {
+  id: string;
+  description: string;
+  rating: number;
+  book_title: string;
+  reviewer: string;
+}
+
+export interface ReviewState {
+  getAllApi: ApiResponse<Array<Review>> | null;
+  overview: { [key: string]: Array<Review> } | null;
+}
 
 export const state: ReviewState = {
   getAllApi: null,
